queue: move DescriptorChain to submodule
Move DescriptorChain/DescriptorChainRwIter to a dedicated file as a
submodule.
